Finish indirection of all Annotation/Document model access
This commit completes the indirection of all access to the Annotation
and Document models through `h.api.storage`. In some places this results
in less than pleasant code, but fear not: in due course all this work
will be reversed.

Unfortunately, in the mean time it is important that we have a single
place in the code where access to the models is mediated so that we can
switch between the ElasticSearch and PostgreSQL models objects on the
basis of a feature flag.

Aspects of this commit worth noting explicitly:

- `storage.annotation_from_dict(data)` constructs an Annotation model
  instance from a data dictionary.
- `Annotation#parent` has been replaced with `Annotation#parent_id`, so
  that the caller is now responsible for fetching the annotation from
  the id.
- `Annotation` and `Document` models are no longer exposed directly from
  `h.models` or from `h.api.models`.