Additional tweaks to the adder enter/exit animation
This animates the vertical position and the opacity of the adder via CSS
transitions when the adder appears and disappears. Additionally, the
color transitions when hovering over the toolbar buttons have also been
animated to soften the changes and avoid a flicker effect.

Unlike the previous commit, the horizontal position is not animated.
The idea is to make it look like a speech/idea bubble popping into focus
above the selected text. The animation is inspired by the Medium editor
toolbar.

 * Animate the opacity/vertical position using CSS transitions.

 * Use 'is-<state>' classnames for states added/removed by JS