Extract overflow state recalculation into a helper class
This makes it possible to write isolated tests for just the overflow
state and content style recalculation aspects of <excerpt> components.

It also simplifies the <excerpt> component tests by making it easier to
reason about when the excerpt's overflow state will have been
recalculated and the directive's state updated accordingly.