Fix error after an automatic incremental rebuild whilst running tests
When running tests in `gulp watch` mode, the initial run passed but
subsequent changes resulted in an error in the `merge-descriptors`
package, related to PhantomJS.

I haven't looked deeply into the problem but it does look like a
familiar issue related to Proxyquire trying to 'call through' to the
original imports in the module. Adding the '@noCallThru' flag in
guest-test.coffee resolves the issue.