Simplify reply notification code
This replaces the supposedly-generic-but-absolutely-completely-not
`generate_notifications` generator function with a `get_notification`, a
function that maps an annotation event to a data structure which can be
used to prepare a reply notification email, 1:1.

This function encapsulates the logic that decides whether or not to send
a reply notification email for a given annotation event, but it is not
responsible for generating or sending the email: that code has moved
into `h.emails.reply_notification` in a prior commit.

One improvement is that we are no longer iterating over all active
subscriptions inside the `get_notification` function in order to find
the one for the relevant user. Instead we use... a database query.

The tests for this package were very hard to modify so I've ended up
essentially rewriting the test suite from scratch, although I have
attempted to ensure that the new test suite covers all cases that the
old one did, as well as several more.