Remove unneeded LoginEvent subscriber
This was added in 3ab65d8 to deal with the fact that we were adding an
opt-in subscriptions table to an already-running service. Why this
wasn't run as a data migration, I can't recall.

Either way, it's not needed any more. Subscription rows are added
for new users on signup (by the lines immediately above), and thereafter
a user can control their subscription using the account control panel.